fix: correctly increment `line_width`
refactor: `split_into_lines` now takes a closure that returns the length for each line

Previously, the width of the text would be appended to `line_width` before it's actually written. This meant that, if text were to be split, text could be pushed into a line without `line_width` being incremented accordingly.
So, any subsequent text's length would be added to 0, then checked, rather than being added to the actual line width.